+++ packages/sessions-sdk-react/src/components/token-list.tsx
@@ -17,12 +17,11 @@ import {
   useTokenAccountData,
 } from "../hooks/use-token-account-data.js";
 
-const MotionGridList = motion.create(GridList<Token>);
+const MotionGridListItem = motion.create(GridListItem<Token>);
 
 type Props = {
   sessionState: EstablishedSessionState;
-  onPressGetTokens: () => void;
-  onPressReceiveTokens: () => void;
+  onPressTransferIn: () => void;
 } & (
   | { onPressToken: (token: Token) => void }
   | { onPressSend: (token: Token) => void }
@@ -30,8 +29,7 @@ type Props = {
 
 export const TokenList = ({
   sessionState,
-  onPressReceiveTokens,
-  onPressGetTokens,
+  onPressTransferIn,
   ...props
 }: Props) => {
   const state = useTokenAccountData(sessionState);
@@ -52,13 +50,11 @@ export const TokenList = ({
           <WalletIcon className={styles.emptyIcon} />
           <span className={styles.message}>Your wallet is empty</span>
           <span className={styles.hints}>
-            <Link onPress={onPressReceiveTokens}>Receive</Link> or{" "}
-            <Link onPress={onPressGetTokens}>Get tokens</Link>
+            <Link onPress={onPressTransferIn}>Transfer USDC to Fogo</Link>
           </span>
         </div>
       ) : (
-        <MotionGridList
-          layoutId="token-list"
+        <GridList
           className={styles.tokenList ?? ""}
           selectionMode="none"
           aria-label="Tokens"
@@ -123,7 +119,9 @@ export const TokenList = ({
               </>
             );
             return (
-              <GridListItem
+              <MotionGridListItem
+                layoutId={mint.toBase58()}
+                layoutScroll
                 textValue={name ?? mint.toBase58()}
                 key={mint.toString()}
                 className={styles.token ?? ""}
@@ -135,10 +133,10 @@ export const TokenList = ({
                 })}
               >
                 {contents}
-              </GridListItem>
+              </MotionGridListItem>
             );
           }}
-        </MotionGridList>
+        </GridList>
       );
     }
     case TokenDataStateType.NotLoaded: