+++ services/paymaster/src/constraint_templates.rs
@@ -108,18 +108,11 @@ impl InstructionConstraint {
     pub fn revoke_session_instruction_constraint() -> InstructionConstraint {
         InstructionConstraint {
             program: fogo_sessions_sdk::session::SESSION_MANAGER_ID,
-            accounts: vec![
-                AccountConstraint {
-                    index: 0,
-                    include: vec![ContextualPubkey::NonFeePayerSigner],
-                    exclude: vec![],
-                },
-                AccountConstraint {
-                    index: 1,
-                    include: vec![ContextualPubkey::Sponsor],
-                    exclude: vec![],
-                },
-            ],
+            accounts: vec![AccountConstraint {
+                index: 0,
+                include: vec![ContextualPubkey::NonFeePayerSigner],
+                exclude: vec![],
+            }],
             data: vec![
                 // instruction = 1 (RevokeSession)
                 DataConstraint {
@@ -129,22 +122,6 @@ impl InstructionConstraint {
                 },
             ],
             required: true,
-        }
-    }
-
-    pub fn create_ata_idempotent_instruction_constraint(required: bool) -> InstructionConstraint {
-        InstructionConstraint {
-            program: spl_associated_token_account::id(),
-            accounts: vec![],
-            data: vec![
-                // instruction = 1 (CreateAssociatedTokenAccountIdempotent)
-                DataConstraint {
-                    start_byte: 0,
-                    data_type: DataType::U8,
-                    constraint: DataConstraintSpecification::EqualTo(vec![DataValue::U8(1)]),
-                },
-            ],
-            required,
         }
     }
 
@@ -198,11 +175,6 @@ impl TransactionVariation {
         TransactionVariation::V1(VariationOrderedInstructionConstraints {
             name: "Session Establishment".to_string(),
             instructions: vec![
-                // Allow for idempotent associated token account creation. For now, we allow up to 4 such optional instructions.
-                InstructionConstraint::create_ata_idempotent_instruction_constraint(false),
-                InstructionConstraint::create_ata_idempotent_instruction_constraint(false),
-                InstructionConstraint::create_ata_idempotent_instruction_constraint(false),
-                InstructionConstraint::create_ata_idempotent_instruction_constraint(false),
                 InstructionConstraint::intent_instruction_constraint(),
                 InstructionConstraint::start_session_instruction_constraint(),
             ],